prevent out-of-bounds reads when using SSE 4.2.
in the SSE4.2 case, ensure the generated static strings are at least
${blob.alignment} characters long. the SSE4.2 _mm* instructions require
data not only to be 16-byte-aligned but also may read full 16 bytes
from the start address.
so far, there is no guarantee that the static strings used for
comparisons are actually in a memory area from which 16 bytes can safely
be read. this is up to the compiler.

motivation for the bugfix: when running an application that is using
llhttp, we are seeing address sanitizer out-of-bounds reads when loading
the static comparison strings into the _mm registers.

this PR is an attempt to fix this, by ensuring that the generated
compare strings are at least 16 bytes long in memory, so that this
memory can safely be read into an _mm register.

caveat: I have just patched the generator code to the best of my
knowledge, but didn't run it yet. but the current code without the
padding definitely has undefined behavior.