fix: speculative execution in wasm
In https://github.com/nodejs/llparse/pull/74 we make sure to increment
position by a constant whenever in a happy path so that the CPU could
proceed through the loop header (`endp - p >= 16`) speculatively instead
of blocking and waiting for vector to scalar to fully resolve.

This doesn't seem to give an obvious performance boost yet, but
potentially might have an effect with the future V8 updates.